small refactoring, added "live" events calendar with events from database
