+++ C_bool.WebApp/Helpers/ImageConverter.cs
@@ -13,13 +13,9 @@ namespace C_bool.WebApp.Helpers
             {
                 if (files.Length > 0)
                 {
-                    //Getting FileName
                     var fileName = Path.GetFileName(files.FileName);
-                    //Getting file Extension
                     var fileExtension = Path.GetExtension(fileName);
-                    // concatenating  FileName + FileExtension
                     var newFileName = String.Concat(Convert.ToString(Guid.NewGuid()), fileExtension);
-
                     string base64String;
                     using (var target = new MemoryStream())
                     {
