+++ C_bool.WebApp/Models/Profiles/UserProfile.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using AutoMapper;
 using C_bool.BLL.DAL.Entities;
 using C_bool.WebApp.Models.Place;
+using C_bool.WebApp.Models.User;
 
 namespace C_bool.WebApp.Models.Profiles
 {
@@ -12,7 +13,7 @@ namespace C_bool.WebApp.Models.Profiles
     {
         public UserProfile()
         {
-            CreateMap<User, UserViewModel>();
+            CreateMap<BLL.DAL.Entities.User, UserViewModel>();
         }
     }
 }
