+++ C_bool.WebApp/Models/User/UserViewModel.cs
@@ -1,9 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using C_bool.BLL.DAL.Entities;
 using C_bool.BLL.Enums;
 
-namespace C_bool.WebApp.Models.Place
+namespace C_bool.WebApp.Models.User
 {
     public class UserViewModel
     {
