+++ C_bool.WebApp/Views/Home/Partials/_EventCalendar.cshtml
@@ -1,0 +1,53 @@
+﻿@using C_bool.WebApp.Models.Place
+@using C_bool.BLL.Enums
+@model IEnumerable<PlaceViewModel>
+@{
+    var events = Model?.SelectMany(x => x.Tasks).Where(x => x.Type == TaskType.Event);
+}
+
+
+<div class="card border-0 shadow">
+    <div class="card-header bg-gray-200 border-bottom">
+        <h3 class="mb-0">Wydarzenia</h3>
+    </div>
+
+    <div class="card-body">
+        @if (events.Count() != 0)
+        {
+            @foreach (var item in events)
+            {
+                <div class="row align-items-center d-block d-sm-flex border-bottom pb-4 mb-4">
+                    <div class="col-auto mb-3 mb-sm-0">
+                        <div class="calendar d-flex"><span class="calendar-month">@item.ValidFrom.ToString("MMM")</span><span class="calendar-day py-2">@item.ValidFrom.ToString("dd")</span></div>
+                    </div>
+                    <div class="col">
+                        <a asp-action="Details" asp-controller="GameTasks" asp-route-gameTaskId="@item.Id">
+                            <h3 class="h5 mb-1">@item.Name</h3>
+                        </a>
+                        <span>Organizowany przez <a href="#">@item.CreatedByName</a></span>
+                        <div class="small fw-bold">Czas: @item.ValidFrom.ToString("dddd, d MMM hh:mm") - @item.ValidThru.ToString("dddd, d MMM hh: mm")</div>
+                        <span class="small fw-bold">
+                            Miejsce:
+                            @if (Model != null)
+                            {
+                                @Model.FirstOrDefault(x => x.Id == item.PlaceId)?.Name
+                            }
+                        </span>
+                    </div>
+                </div>
+            }
+
+        }
+        else
+        {
+            <p>Brak nowych wydarzeń w okolicy w najbliższym czasie, zajrzyj za jakiś czas lub dodaj nowe!</p>
+        }
+
+        </div>
+    @*<div class="card-footer border-top bg-gray-50">
+        <a class="fw-bold d-flex align-items-center justify-content-center" href="../calendar.html">
+            <svg class="icon icon-xs me-2" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path d="M10 12a2 2 0 100-4 2 2 0 000 4z"></path><path fill-rule="evenodd" d="M.458 10C1.732 5.943 5.522 3 10 3s8.268 2.943 9.542 7c-1.274 4.057-5.064 7-9.542 7S1.732 14.057.458 10zM14 10a4 4 0 11-8 0 4 4 0 018 0z" clip-rule="evenodd"></path></svg>
+            Wyświetl wszystkie
+        </a>
+    </div>*@
+</div>
