+++ C_bool.WebApp/Views/NewPlace/SearchNearby.cshtml
@@ -32,48 +32,6 @@
     <div id="map" style="background-color: gainsboro; width: 100%; height: calc(100vh - 270px);"></div>
 </div>
 
-@*<div class="row">
-    <div class="col-md-3">
-        <form asp-action="SearchNearby">
-            <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-            <div class="form-group">
-                <label asp-for="Latitude" class="control-label"></label>
-                <input id="latitudeForm" asp-for="Latitude" class="form-control" onChange="initMap();" />
-                <span asp-validation-for="Latitude" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="Longitude" class="control-label"></label>
-                <input id="longitudeForm" asp-for="Longitude" class="form-control" onchange="initMap();" />
-                <span asp-validation-for="Longitude" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="Radius" class="control-label"></label>
-                <input id="radiusForm" asp-for="Radius" class="form-control" />
-                <span asp-validation-for="Radius" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="Keyword" class="control-label"></label>
-                <input asp-for="Keyword" class="form-control" />
-                <span asp-validation-for="Keyword" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label class="control-label">Kategoria</label>
-                <select asp-for="SelectedType" asp-items="Model.PlaceCategories" class="form-control">
-                    <option selected="selected" value="">Brak kategorii</option>
-                </select>
-            </div>
-            <br />
-            <div class="form-group">
-                <input type="submit" value="Szukaj miejsc" class="btn btn-primary btn-lg" />
-            </div>
-        </form>
-    </div>
-    <div class="col">
-        <div id="map" style="background-color: gainsboro; width: 100%; height: 80vh"></div>
-    </div>
-</div>*@
-
-
 
 @section Scripts {
     @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
