+++ C_bool.WebApp/Views/Shared/_LoginLayout.cshtml
@@ -3,28 +3,6 @@
 @using Microsoft.AspNetCore.Identity
 @using C_bool.BLL.DAL.Entities
 @inject SignInManager<User> SignInManager
-@inject UserManager<User> UserManager
-
-@{ var action = (ViewContext.RouteData.Values["action"] ?? "").ToString().ToLower();
-    var controller = (ViewContext.RouteData.Values["controller"] ?? "").ToString().ToLower();
-    var area = (ViewContext.RouteData.DataTokens["area"] ?? "").ToString().ToLower(); }
-<!--
-
-=========================================================
-* Volt Free - Bootstrap 5 Dashboard
-=========================================================
-
-* Product Page: https://themesberg.com/product/admin-dashboard/volt-bootstrap-5-dashboard
-* Copyright 2021 Themesberg (https://www.themesberg.com)
-* License (https://themesberg.com/licensing)
-
-* Designed and coded by https://themesberg.com
-
-=========================================================
-
-* The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software. Please contact us to request a removal.
-
--->
 
 <!DOCTYPE html>
 <html lang="en">
@@ -145,8 +123,6 @@
     <!-- Simplebar -->
     <script src="/vendor/simplebar/dist/simplebar.min.js"></script>
 
-    <!-- Github buttons -->
-    <script async defer src="https://buttons.github.io/buttons.js"></script>
 
     <!-- Volt JS -->
     <script src="/assets/js/volt.js"></script>
